First pass at addng GENIE info
Added the GENIE header/summary information to the ndflow output

References between GENIE interactions and tracks/segments and the
trajectories seems to have worked.

Applied same unit and axes conversions to the GENIE interaction vertex
to match the trajectories.

Modified how the truth info slices were calculated to be uniform, fix at least one
bug, and to be simpler (in my opinion). The original methods are left in
as commented code.